several changes to try to prevent sigsev faults

+++ TotalCrossVM/src/nm/ui/android/gfx_Graphics_c.h
@@ -292,6 +292,7 @@ void JNICALL Java_totalcross_Launcher4A_nativeInitSize(JNIEnv *env, jobject this
    realAppH = (*env)->CallStaticIntMethod(env, applicationClass, jgetHeight);
    if (lastWindow && lastWindow != window)
    {
+      if (window == null) {debug("window is null. surface is %X. app will likely crash...",surface);}
       destroyEGL();
       initGLES(&screen);
       recreateTextures(false); // now we set the changed flag for all textures
@@ -779,6 +780,7 @@ bool initGLES(ScreenSurface /*screen*/unused)
    EGLint width;
    EGLint height;
 
+   if (!window)                                                             {debug("window is null"); return false;}
    if ((display = eglGetDisplay(EGL_DEFAULT_DISPLAY)) == EGL_NO_DISPLAY)    {debug("eglGetDisplay() returned error %d", eglGetError()); return false;}
    if (!eglInitialize(display, 0, 0))                                       {debug("eglInitialize() returned error %d", eglGetError()); return false;}
    if (!eglChooseConfig(display, attribs, &config, 1, &numConfigs))         {debug("eglChooseConfig() returned error %d", eglGetError()); destroyEGL(); return false;}
